pyproject.toml
[project]
name = "document-analyst-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.11",
    "langchain>=0.3.21",
    "langchain-community>=0.3.20",
    "langchain-experimental>=0.3.4",
    "langchain-openai>=0.3.9",
    "nltk>=3.9.1",
    "pre-commit>=4.2.0",
    "python-dotenv>=1.0.1",
    "spacy>=3.8.5",
    "uvicorn>=0.34.0",
]

[dependency-groups]
dev = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.11.14",
    "aiosignal==1.3.2",
    "altair==5.5.0",
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "asgiref==3.8.1",
    "async-timeout==4.0.3",
    "attrs==25.3.0",
    "backoff==2.2.1",
    "bcrypt==4.3.0",
    "blinker==1.9.0",
    "build==1.2.2.post1",
    "cachetools==5.5.2",
    "certifi==2025.1.31",
    "cffi==1.17.1",
    "charset-normalizer==3.4.1",
    "chroma-hnswlib==0.7.6",
    "chromadb==0.6.3",
    "click==8.1.8",
    "coloredlogs==15.0.1",
    "dataclasses-json==0.6.7",
    "deprecated==1.2.18",
    "distro==1.9.0",
    "durationpy==0.9",
    "exceptiongroup==1.2.2",
    "fastapi==0.115.11",
    "filelock==3.18.0",
    "flatbuffers==25.2.10",
    "frozenlist==1.5.0",
    "fsspec==2025.3.0",
    "gitdb==4.0.12",
    "gitpython==3.1.44",
    "google-auth==2.38.0",
    "googleapis-common-protos==1.69.2",
    "grpcio==1.71.0",
    "h11==0.14.0",
    "httpcore==1.0.7",
    "httptools==0.6.4",
    "httpx==0.28.1",
    "httpx-sse==0.4.0",
    "huggingface-hub==0.29.3",
    "humanfriendly==10.0",
    "idna==3.10",
    "importlib-metadata==8.6.1",
    "importlib-resources==6.5.2",
    "jinja2==3.1.6",
    "jiter==0.9.0",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "jsonschema==4.23.0",
    "jsonschema-specifications==2024.10.1",
    "kubernetes==32.0.1",
    "langchain==0.3.21",
    "langchain-community==0.3.20",
    "langchain-core==0.3.47",
    "langchain-experimental==0.3.4",
    "langchain-openai==0.3.9",
    "langchain-text-splitters==0.3.7",
    "langsmith==0.3.18",
    "lxml==5.3.1",
    "markdown-it-py==3.0.0",
    "markupsafe==3.0.2",
    "marshmallow==3.26.1",
    "mdurl==0.1.2",
    "mmh3==5.1.0",
    "monotonic==1.6",
    "mpmath==1.3.0",
    "multidict==6.2.0",
    "mypy-extensions==1.0.0",
    "narwhals==1.32.0",
    "numpy==2.2.4",
    "oauthlib==3.2.2",
    "onnxruntime==1.21.0",
    "openai==1.68.0",
    "opentelemetry-api==1.31.1",
    "opentelemetry-exporter-otlp-proto-common==1.31.1",
    "opentelemetry-exporter-otlp-proto-grpc==1.31.1",
    "opentelemetry-instrumentation==0.52b1",
    "opentelemetry-instrumentation-asgi==0.52b1",
    "opentelemetry-instrumentation-fastapi==0.52b1",
    "opentelemetry-proto==1.31.1",
    "opentelemetry-sdk==1.31.1",
    "opentelemetry-semantic-conventions==0.52b1",
    "opentelemetry-util-http==0.52b1",
    "orjson==3.10.15",
    "overrides==7.7.0",
    "packaging==24.2",
    "pandas==2.2.3",
    "pillow==11.1.0",
    "posthog==3.21.0",
    "propcache==0.3.0",
    "protobuf==5.29.4",
    "pyarrow==19.0.1",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.1",
    "pycparser==2.22",
    "pydantic==2.10.6",
    "pydantic-core==2.27.2",
    "pydantic-settings==2.8.1",
    "pydeck==0.9.1",
    "pygments==2.19.1",
    "pypdf==5.4.0",
    "pypika==0.48.9",
    "pyproject-hooks==1.2.0",
    "python-dateutil==2.9.0.post0",
    "python-docx==1.1.2",
    "python-dotenv==1.0.1",
    "python-multipart==0.0.20",
    "pytz==2025.2",
    "pyyaml==6.0.2",
    "referencing==0.36.2",
    "regex==2024.11.6",
    "requests==2.32.3",
    "requests-oauthlib==2.0.0",
    "requests-toolbelt==1.0.0",
    "rich==13.9.4",
    "rpds-py==0.23.1",
    "rsa==4.9",
    "shellingham==1.5.4",
    "six==1.17.0",
    "smmap==5.0.2",
    "sniffio==1.3.1",
    "sounddevice==0.5.1",
    "sqlalchemy==2.0.39",
    "starlette==0.46.1",
    "streamlit==1.43.2",
    "sympy==1.13.3",
    "tenacity==9.0.0",
    "tiktoken==0.9.0",
    "tokenizers==0.21.1",
    "toml==0.10.2",
    "tomli==2.2.1",
    "tornado==6.4.2",
    "tqdm==4.67.1",
    "typer==0.15.2",
    "typing-extensions==4.12.2",
    "typing-inspect==0.9.0",
    "tzdata==2025.2",
    "urllib3==2.3.0",
    "uvicorn==0.34.0",
    "uvloop==0.21.0",
    "watchfiles==1.0.4",
    "websocket-client==1.8.0",
    "websockets==15.0.1",
    "wrapt==1.17.2",
    "yarl==1.18.3",
    "zipp==3.21.0",
    "zstandard==0.23.0",
]
